refactor: send email module refactor
