+++ src/user/services/user/user.service.ts
@@ -6,7 +6,7 @@ import {
 import { PrismaService } from 'src/database/prisma.service';
 import { TUser } from 'src/user/ultils/types';
 import { User } from '@prisma/client';
-import { SendEmailService } from 'src/send-email/service/send-email/send-email.service';
+import { SendEmailService } from 'src/send-email/service/send-email.service';
 import { HashService } from 'src/hash/service/hash.service';
 import { LoggerService } from 'src/logger/logger.service';
